added p value
using chi function got chi clearly its 1.19 for first which is ok then like 5 for including the bump in the signal....thats a good sign, we want a higher chi value when the signal bump is included. The import scipy.stats is to calculate a p-value for each the background only and then the including signal hypotehsis.